rework spider job framework(final :v)
