+++ src/main/java/com/swd391/assi2/team2/WebCrawlingApplication.java
@@ -3,10 +3,6 @@ package com.swd391.assi2.team2;
 import com.swd391.assi2.team2.spider.Spider;
 import com.swd391.assi2.team2.spider.SpiderFactory;
 import com.swd391.assi2.team2.spider.job.JobFactory;
-import com.swd391.assi2.team2.spider.job.SpiderJob;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.select.Elements;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
