+++ src/main/java/com/swd391/assi2/team2/spider/job/core/center/MapJob.java
@@ -1,0 +1,15 @@
+package com.swd391.assi2.team2.spider.job.core.center;
+
+import com.swd391.assi2.team2.spider.job.core.SpiderJob;
+import org.jsoup.nodes.Element;
+
+import java.util.ArrayList;
+
+public interface MapJob extends SpiderJob {
+	/**
+	 * @param elements : list need to map to <jobName/>
+	 * @return a new different list with different type of elements
+	 * */
+	ArrayList<Element> map(ArrayList<Element> elements);
+
+}
