+++ src/main/java/com/swd391/assi2/team2/spider/job/imp/Link.java
@@ -1,6 +1,7 @@
-package com.swd391.assi2.team2.spider.job.result;
+package com.swd391.assi2.team2.spider.job.imp;
 
 
+import com.swd391.assi2.team2.spider.job.core.end.OutJob;
 import org.jsoup.nodes.Element;
 
 import java.util.ArrayList;
